sb_loader.exe - v0.2
- Fix a important crash about sb_loader when load a special size sb file

git-svn-id: http://psg-host.svn-tx30.am.freescale.net/svn/psg-host/branches/dev/0004/Apps/sb_loader@337 41b5e5e7-094d-0410-8aea-ef766df42c58

+++ sthiddevice.cpp
@@ -248,7 +248,7 @@ int CStHidDevice::Download(UCHAR* data, ULONGLONG size, CString indent)
 	}
 
 	if( current_pos_in_arr + ulWriteSize > size)
-		ulWriteSize = size - current_pos_in_arr;
+		ulWriteSize = (ULONG)(size - current_pos_in_arr);
 	
 	if( ulWriteSize >= m_Capabilities.OutputReportByteLength)
 		ulWriteSize = m_Capabilities.OutputReportByteLength -1;
